requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Supabase (handles database)
supabase==2.9.1
postgrest==0.17.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pyjwt==2.9.0

# Redis
redis==5.2.0

# Socket.IO
python-socketio==5.11.4
python-engineio==4.10.1

# HTTP Client
httpx==0.27.2

# Email
sendgrid==6.11.0

# Validation (Python 3.13 compatible)
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Task Queue
celery==5.4.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
shortuuid==1.0.13

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.2
